% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{simulate_wrapper}
\alias{simulate_wrapper}
\title{Wrapper SQL Translation}
\usage{
simulate_wrapper()
}
\value{
\itemize{
\item \code{simulate_wrapper()} returns an object of class "WrapperConnection"
that can be used to test SQL generation
}
}
\description{
Wrapper SQL Translation
}
\examples{
library(dplyr, warn.conflicts = FALSE)

lf <- dbplyr::lazy_frame(a = TRUE, b = 1, c = 2, d = "z", con = simulate_wrapper())
lf \%>\% summarise(x = sd(b, na.rm = TRUE))
lf \%>\% summarise(y = cor(b, c), z = cov(b, c))

}
