% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver-dbi.R
\name{wrapper}
\alias{wrapper}
\title{Wrapper DBI implementation}
\usage{
wrapper()
}
\value{
\itemize{
\item \code{wrapper()} returns a DBIDriver subclass
}
}
\description{
Wrapper DBI implementation
}
\examples{
\dontshow{if (has_wrapper_test_flightsql()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
con <- DBI::dbConnect(
  wrapper(),
  wrapper_test_flightsql_uri(),
  username = wrapper_test_flightsql_username(),
  password = wrapper_test_flightsql_password(),
  database_options = c(
    "adbc.flight.sql.client_option.tls_skip_verify" = "true"
  )
)

DBI::dbDisconnect(con)
\dontshow{\}) # examplesIf}
}
