% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wrapper_test_flightsql_uri}
\alias{wrapper_test_flightsql_uri}
\alias{wrapper_test_flightsql_username}
\alias{wrapper_test_flightsql_password}
\alias{has_wrapper_test_flightsql}
\alias{wrapper_test_con}
\title{Wrapper testing parameters}
\usage{
wrapper_test_flightsql_uri()

wrapper_test_flightsql_username()

wrapper_test_flightsql_password()

has_wrapper_test_flightsql()

wrapper_test_con()
}
\value{
\itemize{
\item \code{wrapper_flightsql_test_uri()} returns a URI to a DuckDB-backed
FlightSQL instance. This URI is read from the
R_WRAPPER_FLIGHTSQL_TEST_URI environment variable.
\item \code{wrapper_has_test_flightsql()} returns TRUE if the
R_WRAPPER_FLIGHTSQL_TEST_URI environment variable is set or FALSE
otherwise.
}
}
\description{
Wrapper testing parameters
}
\examples{
\dontshow{if (has_wrapper_test_flightsql()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
has_wrapper_test_flightsql()
wrapper_test_flightsql_uri()
\dontshow{\}) # examplesIf}
}
