% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver-dbi.R
\docType{class}
\name{WrapperConnection-class}
\alias{WrapperConnection-class}
\title{Wrapper S4 DBIConnection}
\description{
Wrapper S4 DBIConnection
}
\section{Slots}{

\describe{
\item{\code{parent_database}}{An ADBC database pointer}

\item{\code{parent_connection}}{An ADBC connection pointer}
}}

