% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver-dbi.R
\docType{class}
\name{WrapperDriver-class}
\alias{WrapperDriver-class}
\alias{WrapperDriver}
\title{Wrapper S4 DBIDriver internals}
\description{
Wrapper S4 DBIDriver internals
}
\section{Slots}{

\describe{
\item{\code{parent_driver}}{An instance of \code{\link[=adbcwrapper]{adbcwrapper()}}}
}}

