% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver-dbi.R
\docType{class}
\name{WrapperResult-class}
\alias{WrapperResult-class}
\title{Wrapper S4 DBIResult}
\description{
Wrapper S4 DBIResult
}
\section{Slots}{

\describe{
\item{\code{reader}}{A nanoarrow_array_stream}
}}

